tasks/prepare_neighborhoods/requirements.txt
functions-framework==3.*
google-cloud-storage==2.*
python-dotenv==1.*
